+++ NEWS.md
@@ -1,8 +1,7 @@
 # mapSpain (development version)
 
 -   Upgrade `leaflet.providersESP.df` to
-    <https://dieghernan.github.io/leaflet-providersESP/> v1.3.1. Fix
-    `"MDT"/"MDT.Relieve"` provider.
+    <https://dieghernan.github.io/leaflet-providersESP/> v1.3.2. 
 
 # mapSpain 0.6.2