+++ NEWS.md
@@ -1,10 +1,12 @@
 # mapSpain (development version)
 
 -   Upgrade `leaflet.providersESP.df` to
-    <https://dieghernan.github.io/leaflet-providersESP/> v1.3.2. 
+    <https://dieghernan.github.io/leaflet-providersESP/> v1.3.2.
 
--  Changes on how to package manages tiles providers:
-  - `leaflet.providersESP.df` is superseded in favor of `esp_tiles_providers`.
+-   Changes on how to package manages tiles providers:
+
+    -   `leaflet.providersESP.df` is superseded in favor of
+        `esp_tiles_providers`.
 
 # mapSpain 0.6.2